Simplify tag extraction in attach
